+++ R/print-methods.R
@@ -569,3 +569,18 @@ print.check_collinearity <- function(x, ...) {
     print.data.frame(x[high_corr, ], row.names = FALSE)
   }
 }
+
+
+
+#' @importFrom insight format_table print_color
+#' @export
+print.performance_lrt <- function(x, digits = 3, ...) {
+  insight::print_color("# Likelihood-Ratio-Test for Model Comparison\n\n", "blue")
+  cat(insight::format_table(x, digits = digits))
+
+  if (any(x$p < .05) && sum(x$p < .05, na.rm = TRUE) <= 1) {
+    best <- which(x$p < .05)
+    if (length(best) == 0) best <- 1
+    insight::print_color(sprintf("\nModel '%s' seems to have the best model fit.\n", x$Model[best]), "yellow")
+  }
+}